Refactor HomePage into components
