+++ src/components/GlobalStyles.jsx
@@ -1,0 +1,258 @@
+import React from "react";
+
+const GlobalStyles = React.memo(() => (
+  <style>{`
+    /* Ensure html and body take full viewport height and remove default margins */
+    html, body {
+        height: 100%;
+        margin: 0;
+        padding: 0;
+        /* Custom cursor: a yellow mouse pointer */
+        cursor: url("data:image/svg+xml;utf8,<svg%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20width%3D%2224%22%20height%3D%2224%22%20viewBox%3D%220%200%2024%2024%22%20fill%3D%22none%22%20stroke%3D%22%23FACC15%22%20stroke-width%3D%222%22%20stroke-linecap%3D%22round%22%20stroke-linejoin%3D%22round%22%3E%0A%20%20%3Cpath%20d%3D%22M3%203l7.07%2017.07%202.51-7.07%207.07-2.51L3%203z%22%2F%3E%0A%3C%2Fsvg%3E"), auto;
+    }
+
+    /* Apply background to a top-level container within the React app */
+    .app-container {
+        min-height: 100vh; /* Ensure it always covers the viewport height */
+        display: flex;
+        flex-direction: column;
+        background-color: #0D0D0D;
+        background-image:
+            linear-gradient(to right, rgba(255,255,255,0.03) 1px, transparent 1px), /* Reverted opacity to original */
+            linear-gradient(to bottom, rgba(255,255,255,0.03) 1px, transparent 1px); /* Reverted opacity to original */
+        background-size: 24px 24px;
+    }
+
+    /* Main content area should grow to fill available space */
+    main {
+        flex-grow: 1;
+    }
+
+    /* Custom scrollbar for a more modern feel */
+    ::-webkit-scrollbar {
+        width: 8px;
+    }
+    ::-webkit-scrollbar-track {
+        background: #111111;
+    }
+    ::-webkit-scrollbar-thumb {
+        background: #FACC15;
+        border-radius: 4px;
+    }
+    ::-webkit-scrollbar-thumb:hover {
+        background: #EAB308;
+    }
+    html {
+        scroll-behavior: smooth;
+    }
+    .carousel-container {
+        overflow: hidden;
+        position: relative;
+    }
+    .carousel-track {
+        display: flex;
+        transition: transform 0.5s ease-in-out;
+    }
+    .carousel-slide {
+        min-width: 100%;
+        box-sizing: border-box;
+    }
+    @media (min-width: 768px) {
+        .carousel-slide {
+            min-width: 50%;
+        }
+    }
+    @media (min-width: 1024px) {
+        .carousel-slide {
+            min-width: 33.3333%;
+        }
+    }
+    .feature-card, .grid-card, .info-card {
+        transition: transform 0.3s ease, box-shadow 0.3s ease, border-color 0.3s ease-in-out, background-color 0.3s ease;
+        display: block;
+        border: 2px solid transparent;
+        box-shadow: 0 10px 15px -3px rgb(0 0 0 / 0.1), 0 4px 6px -4px rgb(0 0 0 / 0.1);
+    }
+    .feature-card:hover, .grid-card:hover, .info-card:hover {
+        transform: translateY(-8px);
+        box-shadow: 0 10px 20px rgba(250, 204, 21, 0.15);
+        border-color: #FACC15;
+    }
+    .grid-card:hover .icon-container i, .feature-card:hover .icon-container i, .info-card:hover .icon-container i,
+    .grid-card:hover .icon-container svg, .feature-card:hover .icon-container svg, .info-card:hover .icon-container svg {
+        transform: scale(1.1) rotate(5deg);
+        color: #EAB308;
+    }
+    .icon-container i, .icon-container svg {
+        transition: transform 0.3s ease, color 0.3s ease;
+    }
+    .feature-card, .grid-card, .info-card {
+        background-color: rgba(30, 30, 30, 0.9);
+        color: #cccccc;
+        height: 100%;
+    }
+    .info-card h3 {
+        color: #f5f5f5;
+    }
+    .info-card p {
+        color: #aaaaaa;
+    }
+    .feature-card h3, .grid-card h3 {
+        color: #f5f5f5;
+    }
+    .feature-card p, .grid-card p {
+        color: #aaaaaa;
+    }
+    .feature-card a {
+        color: #FACC15;
+    }
+    .feature-card a:hover {
+        color: #EAB308;
+    }
+    .continuous-carousel-viewport {
+        overflow: hidden;
+        width: 100%;
+        -webkit-mask-image: linear-gradient(to right, transparent, black 15%, black 85%, transparent);
+        mask-image: linear-gradient(to right, transparent, black 15%, black 85%, transparent);
+    }
+    .continuous-carousel-track {
+        display: flex;
+        width: fit-content;
+        animation: scrollLeftContinuous 60s linear infinite;
+    }
+    .continuous-carousel-container:hover .continuous-carousel-track {
+        animation-play-state: paused;
+    }
+    .continuous-carousel-card {
+        flex-shrink: 0;
+        width: 200px;
+        height: 100px;
+        background-color: #222222;
+        color: #cccccc;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+        text-align: center;
+        margin: 0 0.5rem;
+        padding: 0.5rem;
+        border-radius: 0.5rem;
+        transition: background-color 0.3s ease;
+        font-weight: 600;
+    }
+    .continuous-carousel-card i, .continuous-carousel-card svg {
+        margin-bottom: 0.5rem;
+    }
+    .continuous-carousel-card:hover {
+        background-color: #333333;
+    }
+    .continuous-carousel-row-2 {
+        margin-left: 3rem;
+    }
+    @keyframes scrollLeftContinuous {
+        0% {
+            transform: translateX(0%);
+        }
+        100% {
+            transform: translateX(-50%);
+        }
+    }
+    .footer-link-column h4 {
+        margin-bottom: 0.75rem;
+        font-size: 0.875rem;
+        font-weight: 600;
+        color: #f5f5f5;
+        text-transform: uppercase;
+    }
+    .footer-link-column ul li a {
+        display: inline-block;
+        padding: 0.25rem 0;
+        font-size: 0.875rem;
+        color: #cccccc;
+        transition: color 0.2s ease-in-out;
+    }
+    .footer-link-column ul li a:hover {
+        color: #FACC15;
+    }
+    .nav-link {
+        color: #cccccc;
+        transition: color 0.3s ease-in-out;
+    }
+    .nav-link:hover {
+        color: #FACC15;
+    }
+    #mobileMenu a.nav-link-mobile {
+        color: #cccccc;
+        transition: color 0.3s ease-in-out, background-color 0.3s ease-in-out;
+    }
+    #mobileMenu a.nav-link-mobile:hover {
+        color: #FACC15;
+        background-color: #222222;
+    }
+    .animated-cta-button {
+        border-width: 2px;
+        border-color: transparent;
+        box-shadow: 0 10px 15px -3px rgb(0 0 0 / 0.1), 0 4px 6px -4px rgb(0 0 0 / 0.1);
+        transition: box-shadow 0.3s ease, border-color 0.3s ease-in-out, background-color 0.3s ease;
+    }
+    .animated-cta-button:hover {
+        border-color: #FACC15;
+        box-shadow: 0 10px 20px rgba(250, 204, 21, 0.15);
+    }
+    .eat-cheese-dynamic-text {
+        display: inline-block;
+        transition: opacity 0.3s ease-in-out, transform 0.3s ease-in-out;
+        opacity: 1;
+    }
+    .eat-cheese-dynamic-text.fade-out {
+        opacity: 0;
+        transform: translateY(10px);
+    }
+    .eat-cheese-dynamic-text.fade-in {
+        opacity: 1;
+        transform: translateY(0);
+    }
+    .eat-cheese-icon {
+        transition: transform 0.2s ease-in-out, color 0.2s ease-in-out, border-color 0.2s ease-in-out;
+        cursor: pointer;
+        padding: 0.75rem;
+        border: 2px solid transparent;
+        border-radius: 0.5rem;
+        color: #CCCCCC;
+    }
+    .eat-cheese-icon:hover, .eat-cheese-icon.active {
+        transform: scale(1.15);
+        color: #FACC15;
+        border-color: #FACC15;
+    }
+    .eat-cheese-icon svg {
+        fill: none;
+        stroke: currentColor;
+        stroke-width: 2;
+    }
+    .idea-card {
+        background-color: rgba(30, 30, 30, 0.9);
+        border: 2px solid #FACC15;
+        min-height: 150px;
+    }
+    .loader {
+        border: 4px solid #333333;
+        border-top: 4px solid #FACC15;
+        border-radius: 50%;
+        width: 40px;
+        height: 40px;
+        animation: spin 1s linear infinite;
+        margin: 20px auto;
+    }
+    @keyframes spin {
+        0% { transform: rotate(0deg); }
+        100% { transform: rotate(360deg); }
+    }
+    .bg-theme-translucent-dark {
+      background-color: rgba(17, 17, 17, 0.9);
+    }
+  `}</style>
+));
+GlobalStyles.displayName = "GlobalStyles";
+
+export default GlobalStyles;
